+++ Laboratorium6/Laboratorium6/Globals.h
@@ -3,5 +3,5 @@ class Vector;
 std::ostream& operator<<(std::ostream&, Point2d&);
 std::ostream& operator<<(std::ostream&, Point3d&);
 std::ostream& operator<<(std::ostream&, Vector&);
-float operator!(const Vector&);
+double operator!(const Vector&);
 Vector operator*(float, Vector&);
